SIMSBIOHUB-192: PR changes
